+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - whenever user navigates to the main page, it will automatically set the user's org to be the first of the fetched orgs. otherwise it will stay on the create orgs screen
 - Moved organizations button into the user dropdown
 - Updated project creation dates to display in a more understandable manner
+- Goals and Tasks now support acceptance criteria
 
 ### Bug Fixes